+++ src/main/java/com/smoorbuilderserver/repository/ClassDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.ClassDescription;
 
 @Repository
-public interface ClassDescriptionRepository extends CrudRepository<ClassDescription, Long> {
+public interface ClassDescriptionRepository extends CrudRepository<ClassDescription, BigInteger> {
 	
 	public ClassDescription findByClassName(String generalClassName);
 	
-	public ClassDescription findByid(Integer id);
+	public ClassDescription findByid(BigInteger id);
 	
 	public List<ClassDescription> findDistinctClassDescriptionByClassName(String generalClassName);
 }
