+++ src/main/java/com/smoorbuilderserver/repository/RaceDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.RaceDescription;
 
 @Repository
-public interface RaceDescriptionRepository extends CrudRepository<RaceDescription, Long> {
+public interface RaceDescriptionRepository extends CrudRepository<RaceDescription, BigInteger> {
 	
 	public RaceDescription findByRaceName(String raceName);
 	
-	public RaceDescription findByid(Integer id);
+	public RaceDescription findByid(BigInteger id);
 	
 	public List<RaceDescription> findDistinctRaceDescriptionByRaceName(String raceName);
 }
